# api/openapi.yaml
openapi: 3.0.3
info:
  title: authsvc API
  version: 0.3.0
  description: Plug-and-play authentication and authorization service
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
    BootstrapSecret:
      type: http
      scheme: bearer
      description: Bootstrap secret as Bearer token for initial admin setup
...
paths:
...
  /v1/clients:
    post:
      summary: Create OAuth client (admin)
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      responses:
        "200": { description: Client created }
    get:
      summary: List OAuth clients (admin)
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      responses:
        "200": { description: Client list }
...
  /v1/keys/rotate:
    post:
      summary: Rotate JWT signing keys
      security:
        - BearerAuth: []
      responses:
        "200": { description: New kid }
...
  /v1/compliance/status:
    get:
      summary: Compliance posture (admin)
      security:
        - BearerAuth: []
      responses:
        "200": { description: Compliance status }
...
  /v1/clients/{id}:
    delete:
      summary: Delete OAuth client (admin)
      security:
        - BearerAuth: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "204": { description: Deleted }
...
  /v1/saml/service-providers:
    get:
      summary: List SAML IdP service providers (admin)
      security:
        - BearerAuth: []
      responses:
        "200": { description: Service provider list }
    post:
      summary: Register SAML service provider (admin)
      security:
        - BearerAuth: []
      responses:
        "200": { description: Service provider created }
  /v1/saml/service-providers/{id}:
    delete:
      summary: Delete SAML service provider (admin)
      security:
        - BearerAuth: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "204": { description: Deleted }
...
  /admin/session:
    post:
      summary: Create admin console session (httpOnly cookie)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token: { type: string }
      responses:
        "204": { description: Session created }
    get:
      summary: Admin session status
      responses:
        "200": { description: Session status }
    delete:
      summary: Destroy admin session
      responses:
        "204": { description: Session cleared }
...